Part 1 of Aki Koskinen linux patch: 1) The biggest change is the addition of shared/linux/LinuxMain.cpp which initializes the graphics system for Linux desktops. 2) A new preprocessor definition RT_USE_SDL_AUDIO is introduced and used in shared/Audio/AudioManagerSDL* to possibly enable the SDL audio system on other than WebOS systems too. 3) In shared/Audio/AudioManager.h the AudioHandle type is now defined as uintptr_t. Pointers to various objects are cast to AudioHandle throughout the code. The previous definition uint32 is too short on 64 bit systems and might cause precision loss. uintptr_t should be the same size that void* type is on all systems. (This change isn't Linux specific in any sense but because of the different size of pointer types in 32 and 64 bit architectures – regardless of the OS. I just happen to have a 64 bit system...) 4) Other minor changes include adding the std namespace and some letter case changes in various places to get stuff compiled.
Also some misc unrelated stuff from Seth:
* Added Surface::SetCreateMipMapsIfNeeded(), must use this if you want mipmap chains created (if they the file didn't include them), without this, functionality slightly changed and broke existing stuff, can't have that
* Cleaned up warnings in RenderScissorComponent.cpp
* Cleaned up warnings in RenderUtils.cpp

+++ shared/Audio/AudioManager.h
@@ -21,7 +21,7 @@
 #define AudioManager_h__
 
 
-typedef uint32 AudioHandle;
+typedef uintptr_t AudioHandle;
 
 #define AUDIO_HANDLE_BLANK 0
